+++ .gitignore
@@ -12,6 +12,8 @@ dist
 dist-ssr
 *.local
 
+#just some files
+src/saveBody.jsx
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
