+++ src/components/About.jsx
@@ -1,0 +1,8 @@
+const About = () => {
+	return (
+		<div>
+		<p>this is About Page</p>
+		</div>
+	)
+}
+export default About
