+++ src/components/Header.jsx
@@ -1,10 +1,11 @@
+import {Link} from 'react-router-dom';
 const Header = () => {
 	return(
 		<div className="header">
 		<div className="bar">
-		<div className="logo">LightUI</div>
+		<div className="logo"><Link to='/'>LightUI</Link></div>
 		<div className="otherele">
-		<p>About Us</p>
+		<p><Link to='/about'>About Us</Link></p>
 		<p>Contact Us</p>
 		</div>
 		</div>
